feat: unit tests

+++ ifood-vemsertech-backend/1090-BE-JS-009-TESTES-AUTOMATIZADOS/home-work/final-test/src/controllers/books/delete.ts
@@ -12,6 +12,12 @@ export class DeleteBooksController {
     const { id } = req.params
 
     try {
+      const book = await this.booksRepository.getById(id)
+      if(!book){
+        res.status(409).json({ message: 'any book with the id provided was founded' })
+        return
+      }
+
       await this.booksRepository.delete(id)
       res.status(200).send()
       return
