+++ ifood-vemsertech-backend/1090-BE-JS-009-TESTES-AUTOMATIZADOS/home-work/final-test/src/controllers/books_rental/delete.ts
@@ -12,6 +12,12 @@ export class DeleteBooksRentalController {
     const { id } = req.params
 
     try {
+      const booksRental = await this.booksRentalRepository.getById(id)
+      if(!booksRental){
+        res.status(409).json({ message: 'any book rental with the id provided was founded' })
+        return
+      }
+
       await this.booksRentalRepository.delete(id)
       res.status(200).send()
       return
